Fixed user's common errors
1) Set tankCapacity to double, 2) Checked initial createProfile’s path
history for null situation

+++ obdEnergy/src/main/java/com/example/obdenergy/obdenergy/Utilities/Calculations.java
@@ -22,7 +22,7 @@ public class Calculations {
         DecimalFormat df = new DecimalFormat("#.00");
 
         Console.log("Calculations - initfuel, final fuel, tankcapacity"+initFuel+" "+finalFuel+" "+tankCapacity);
-        int tankCapacityNum = Integer.parseInt(tankCapacity);
+        double tankCapacityNum = Double.parseDouble(tankCapacity);
 
         double fuelVal = finalFuel - initFuel;
         double percentage = (fuelVal/255);
